Fixed script

+++ scripts/kill-redis.sh
@@ -1,7 +1,5 @@
 #!/bin/sh
 
+set -x
 pids=`$(ps aux | grep 'redis' | grep -v grep | awk '{print $2}')`
-if "$pids" != "";
-  echo "Killing existing redis instances"
-  kill $pids
-fi
+test "$pids" != "" && kill $pids
